backend/requirements.txt
# Core dependencies
psutil==5.9.8
pyyaml==6.0.1

# Testing
pytest==7.4.4
pytest-asyncio==0.23.4
pytest-cov==4.1.0

# Code quality
black==24.1.1
flake8==7.0.0
mypy==1.8.0

# Type stubs
types-PyYAML==6.0.12.12
